Moved LoinGovHelpText out of the Module component.
